Refactor, fix, and streamline test cases
I moved the list of registers and constants into the TestASM
module, and then reduced the list unless TEST_EXHAUSTIVE
is set.  Same with the constants.  This greatly reduces the
time required for the testing.

Then, I rewrote all the iterations over memory address options
so that they use a common block of code and can be reduces or
expanded more easily.  (also written more easily)