New data_strtable method, and populate label into empty scalar-ref
This enables a new simpler and more efficient way to build a string
table:

    ->lea(RSI, [RIP => \$str{$msg}])
    ...
    ->data_strtable(\%str)

for any string $msg, you can automatically create a label into that
field of the hash, and gain automatic de-duplication of string
constants.  Then when building the string table, it can automatically
resolve all the labels at once while also overlapping strings on
eachother when one string is a subset of another.