Overhaul of register numbers and encoding system
The code was getting repetitive and confusing.  I decided that
using distinct register numbers, each with flags about what
conditions they trigger, would be the easiest way to sort out
the mess and reduce the duplication.

It is better now, but would probably still be better to use some
sort of master table of instructions and a single general-purpose
encoding function that can be overridden for the special cases.