Fix more edge cases for 'append': RIP-relative and unanchored labels
The RIP-relative addressing was using objects which had no access
from the _unresolved list, so they would still be referencing the
old encoder after appended to a new one.  I fixed that by giving
rip-relative access to the instruction, and upgrading that to an
object as well, now called 'LazyEncode'.